user_service/requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
pytest==7.4.2
python-dotenv==1.0.0
passlib==1.7.4
pydantic==2.4.2
python-jose[cryptography]==3.3.0
termcolor==2.3.0
pymongo==4.6.1
motor==3.3.2
PyJWT==2.8.0
